+++ config/urls.py
@@ -23,7 +23,7 @@ from wies.core.views import AssignmentDetailView, ColleagueDetailView
 from wies.core.views import PlacementListView
 from wies.core.views import MinistryDetailView
 from wies.core.views import admin_db, login, no_access, logout, auth
-from wies.core.views import UserListView, user_create, user_edit, user_delete, user_import_csv
+from wies.core.views import UserListView, user_create, user_edit, user_delete, user_import_csv, widget_email_row
 from wies.core.views import placement_import_csv
 
 urlpatterns = [
@@ -45,4 +45,5 @@ urlpatterns = [
     path('users/<int:pk>/edit/', user_edit, name='user-edit'),
     path('users/<int:pk>/delete/', user_delete, name='user-delete'),
     path('users/import/', user_import_csv, name='user-import-csv'),
+    path('widgets/email-row/', widget_email_row, name='widget-email-row'),
 ]