+++ wies/core/models.py
@@ -47,7 +47,19 @@ class Skill(models.Model):
         return self.name
 
 class User(AbstractUser):
+    email = models.EmailField(unique=True)
     brand = models.ForeignKey('Brand', models.SET_NULL, null=True, blank=True, related_name='users')
+
+
+class EmailAlias(models.Model):
+    user = models.ForeignKey(User, on_delete=models.CASCADE, related_name='email_aliases')
+    email = models.EmailField(unique=True)
+
+    class Meta:
+        verbose_name_plural = 'email aliases'
+
+    def __str__(self):
+        return self.email
 
 
 class Colleague(models.Model):